t/data/parsing/temma-flow-1.dat

#data
<t:wait><!---->
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <temma wait>
|     <!--  -->

#data
<base><t:wait><link>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <base>
|     <temma wait>
|     <link>

#data
<p><t:wait>abc
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <p>
|       <temma wait>
|       "abc"

#data
<p><t:attr><t:wait><t:attr>ab
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <p>
|       <temma attr>
|       <temma wait>
|       <temma attr>
|       "ab"

#data
<t:if/><!---->
#errors
#document
| <!DOCTYPE html>
| <html>
|   <temma if>
|   <!--  -->

#data
<base><t:if/><!---->
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <base>
|     <temma if>
|     <!--  -->

#data
a<t:if/><!---->
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     "a"
|     <temma if>
|     <!--  -->

#data
a<t:if>xyx</t:if>xy
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     "a"
|     <temma if>
|       "xyx"
|     "xy"

#data
<t:if><t:else>a</t:if>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <temma if>
|     <temma else>
|     "a"

#data
<head><t:if><t:else>a</t:if>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <temma if>
|       <temma else>
|       "a"

#data
<head><t:if>a<t:elsif>b<t:else>c</t:if>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <temma if>
|       "a"
|       <temma elsif>
|       "b"
|       <temma else>
|       "c"

#data
<t:else>a
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <temma else>
|     "a"

#data
<t:elsif>a
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <temma elsif>
|     "a"

#data
b<t:if x=1>xx<p>a<t:else>aa</t:if>y
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     "b"
|     <temma if>
|       x="1"
|       "xx"
|       <p>
|         "a"
|       <temma else>
|       "aa"
|     "y"

#data
b<t:if x=1>xx<p>a<t:elsif>aa</t:if>y
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     "b"
|     <temma if>
|       x="1"
|       "xx"
|       <p>
|         "a"
|       <temma elsif>
|       "aa"
|     "y"

#data
foo<t:if><p><q><t:else>aa</t:if>bb
#errors
1;24;m;not closed;;q
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     "foo"
|     <temma if>
|       <p>
|         <q>
|       <temma else>
|       "aa"
|     "bb"

#data
foo<t:if><p><q><t:elsif>aa</t:if>bb
#errors
1;25;m;not closed;;q
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     "foo"
|     <temma if>
|       <p>
|         <q>
|       <temma elsif>
|       "aa"
|     "bb"

#data
<head><t:if><t:if><t:else>a</t:if>b</t:if>c
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <temma if>
|       <temma if>
|         <temma else>
|         "a"
|       "b"
|   <body>
|     "c"

#data
<t:call><!---->
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <temma call>
|     <!--  -->

#data
a<t:call><!---->
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     "a"
|     <temma call>
|     <!--  -->

#data
<t:for>hoge</t:for><link>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <temma for>
|     "hoge"
|   <head>
|     <link>

#data
<meta><t:for>hoge</t:for><link>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <meta>
|     <temma for>
|       "hoge"
|     <link>

#data
a<t:for>x</t:for>b
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     "a"
|     <temma for>
|       "x"
|     "b"

#data
<t:sep><!---->
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <temma sep>
|     <!--  -->

#data
<head><t:for><!----><t:sep><!----></t:for>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <temma for>
|       <!--  -->
|       <temma sep>
|       <!--  -->

#data
<head><t:for><t:sep><t:sep><t:sep></t:for>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <temma for>
|       <temma sep>
|       <temma sep>
|       <temma sep>

#data
<head><t:for><p>a<t:sep>b</t:for>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <temma for>
|       <p>
|         "a"
|       <temma sep>
|       "b"

#data
<head><t:for><div>a<t:sep>b</t:for>
#errors
1;27;m;not closed;;div
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <temma for>
|       <div>
|         "a"
|       <temma sep>
|       "b"

#data
<head><t:for><t:for><t:sep>a</t:for>b</t:for>c
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <temma for>
|       <temma for>
|         <temma sep>
|         "a"
|       "b"
|   <body>
|     "c"

#data
<ul><li>a</li><t:if><li>b</li></t:if><li>c</li></ul>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <ul>
|       <li>
|         "a"
|       <temma if>
|         <li>
|           "b"
|       <li>
|         "c"

#data
<ul><li>a</li><t:for><li>b</li></t:for><li>c</li></ul>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <ul>
|       <li>
|         "a"
|       <temma for>
|         <li>
|           "b"
|       <li>
|         "c"

#data
<ul><li>a</li><t:try><li>b</li></t:try><li>c</li></ul>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <ul>
|       <li>
|         "a"
|       <temma try>
|         <li>
|           "b"
|       <li>
|         "c"

#data
<ul><li>a<t:if><li>b</t:if><li>c</ul>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <ul>
|       <li>
|         "a"
|         <temma if>
|           <li>
|             "b"
|       <li>
|         "c"

#data
<dl><dt>a</dt><t:if><dt>b</dt></t:if><dt>c</dt></dl>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <dl>
|       <dt>
|         "a"
|       <temma if>
|         <dt>
|           "b"
|       <dt>
|         "c"

#data
<dl><dt>a</dt><t:for><dt>b</dt></t:for><dt>c</dt></dl>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <dl>
|       <dt>
|         "a"
|       <temma for>
|         <dt>
|           "b"
|       <dt>
|         "c"

#data
<dl><dt>a</dt><t:try><dt>b</dt></t:try><dt>c</dt></dl>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <dl>
|       <dt>
|         "a"
|       <temma try>
|         <dt>
|           "b"
|       <dt>
|         "c"

#data
<dl><dd>a</dd><t:if><dd>b</dd></t:if><dd>c</dd></dl>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <dl>
|       <dd>
|         "a"
|       <temma if>
|         <dd>
|           "b"
|       <dd>
|         "c"

#data
<dl><dd>a</dd><t:for><dd>b</dd></t:for><dd>c</dd></dl>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <dl>
|       <dd>
|         "a"
|       <temma for>
|         <dd>
|           "b"
|       <dd>
|         "c"

#data
<dl><dd>a</dd><t:try><dd>b</dd></t:try><dd>c</dd></dl>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <dl>
|       <dd>
|         "a"
|       <temma try>
|         <dd>
|           "b"
|       <dd>
|         "c"

#data
<table><tr><!----><t:if><tr><!----></t:if><tr><!----></table><!---->
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <!--  -->
|           <temma if>
|             <tr>
|               <!--  -->
|         <tr>
|           <!--  -->
|     <!--  -->

#data
<table><tr><!----></tr><t:if><tr><!----></t:if><tr><!----></table><!---->
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <!--  -->
|         <temma if>
|           <tr>
|             <!--  -->
|         <tr>
|           <!--  -->
|     <!--  -->

#data
<table><tr><th><!----><t:if><th><!----></t:if><th><!----></table><!---->
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <th>
|             <!--  -->
|             <temma if>
|               <th>
|                 <!--  -->
|           <th>
|             <!--  -->
|     <!--  -->

#data
<table><tr><td><!----><t:if><td><!----></t:if><td><!----></table><!---->
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <!--  -->
|             <temma if>
|               <td>
|                 <!--  -->
|           <td>
|             <!--  -->
|     <!--  -->

#data
<table><thead><!----><t:if><thead><!----></t:if><thead><!----></table><!---->
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <table>
|       <thead>
|         <!--  -->
|         <temma if>
|           <thead>
|             <!--  -->
|       <thead>
|         <!--  -->
|     <!--  -->

#data
<table><tbody><!----><t:if><tbody><!----></t:if><tbody><!----></table><!---->
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <table>
|       <tbody>
|         <!--  -->
|         <temma if>
|           <tbody>
|             <!--  -->
|       <tbody>
|         <!--  -->
|     <!--  -->

#data
<table><tfoot><!----><t:if><tfoot><!----></t:if><tfoot><!----></table><!---->
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <table>
|       <tfoot>
|         <!--  -->
|         <temma if>
|           <tfoot>
|             <!--  -->
|       <tfoot>
|         <!--  -->
|     <!--  -->

#data
<table><tr><!----><t:for><tr><!----></t:for><tr><!----></table><!---->
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <!--  -->
|           <temma for>
|             <tr>
|               <!--  -->
|         <tr>
|           <!--  -->
|     <!--  -->

#data
<table><tr><!----></tr><t:for><tr><!----></t:for><tr><!----></table><!---->
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <!--  -->
|         <temma for>
|           <tr>
|             <!--  -->
|         <tr>
|           <!--  -->
|     <!--  -->

#data
<table><tr><th><!----><t:for><th><!----></t:for><th><!----></table><!---->
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <th>
|             <!--  -->
|             <temma for>
|               <th>
|                 <!--  -->
|           <th>
|             <!--  -->
|     <!--  -->

#data
<table><tr><td><!----><t:for><td><!----></t:for><td><!----></table><!---->
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <!--  -->
|             <temma for>
|               <td>
|                 <!--  -->
|           <td>
|             <!--  -->
|     <!--  -->

#data
<table><thead><!----><t:for><thead><!----></t:for><thead><!----></table><!---->
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <table>
|       <thead>
|         <!--  -->
|         <temma for>
|           <thead>
|             <!--  -->
|       <thead>
|         <!--  -->
|     <!--  -->

#data
<table><tbody><!----><t:for><tbody><!----></t:for><tbody><!----></table><!---->
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <table>
|       <tbody>
|         <!--  -->
|         <temma for>
|           <tbody>
|             <!--  -->
|       <tbody>
|         <!--  -->
|     <!--  -->

#data
<table><tfoot><!----><t:for><tfoot><!----></t:for><tfoot><!----></table><!---->
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <table>
|       <tfoot>
|         <!--  -->
|         <temma for>
|           <tfoot>
|             <!--  -->
|       <tfoot>
|         <!--  -->
|     <!--  -->

#data
<table><tr><!----><t:try><tr><!----></t:try><tr><!----></table><!---->
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <!--  -->
|           <temma try>
|             <tr>
|               <!--  -->
|         <tr>
|           <!--  -->
|     <!--  -->

#data
<table><tr><!----></tr><t:try><tr><!----></t:try><tr><!----></table><!---->
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <!--  -->
|         <temma try>
|           <tr>
|             <!--  -->
|         <tr>
|           <!--  -->
|     <!--  -->

#data
<table><tr><th><!----><t:try><th><!----></t:try><th><!----></table><!---->
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <th>
|             <!--  -->
|             <temma try>
|               <th>
|                 <!--  -->
|           <th>
|             <!--  -->
|     <!--  -->

#data
<table><tr><td><!----><t:try><td><!----></t:try><td><!----></table><!---->
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <!--  -->
|             <temma try>
|               <td>
|                 <!--  -->
|           <td>
|             <!--  -->
|     <!--  -->

#data
<table><thead><!----><t:try><thead><!----></t:try><thead><!----></table><!---->
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <table>
|       <thead>
|         <!--  -->
|         <temma try>
|           <thead>
|             <!--  -->
|       <thead>
|         <!--  -->
|     <!--  -->

#data
<table><tbody><!----><t:try><tbody><!----></t:try><tbody><!----></table><!---->
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <table>
|       <tbody>
|         <!--  -->
|         <temma try>
|           <tbody>
|             <!--  -->
|       <tbody>
|         <!--  -->
|     <!--  -->

#data
<table><tfoot><!----><t:try><tfoot><!----></t:try><tfoot><!----></table><!---->
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <table>
|       <tfoot>
|         <!--  -->
|         <temma try>
|           <tfoot>
|             <!--  -->
|       <tfoot>
|         <!--  -->
|     <!--  -->

#data
<select><option>a</option><t:for><option>b</t:for></select>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <select>
|       <option>
|         "a"
|       <temma for>
|         <option>
|           "b"

#data
<select><optgroup>a</optgroup><t:for><optgroup>b</t:for></select>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <select>
|       <optgroup>
|         "a"
|       <temma for>
|         <optgroup>
|           "b"

#data
<head><t:for>hoge<t:next>aaa</t:for>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <temma for>
|       "hoge"
|       <temma next>
|       "aaa"

#data
<head><t:for>hoge<t:last>aaa</t:for>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <temma for>
|       "hoge"
|       <temma last>
|       "aaa"

#data
<head><t:for>ho<p>ge<t:next>aaa</t:for>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <temma for>
|       "ho"
|       <p>
|         "ge"
|         <temma next>
|         "aaa"

#data
<head><t:for>ho<p>ge<t:last>aaa</t:for>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <temma for>
|       "ho"
|       <p>
|         "ge"
|         <temma last>
|         "aaa"

#data
a<p><t:attr><t:for><t:attr><t:if><t:last></t:if></t:for>x
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     "a"
|     <p>
|       <temma attr>
|       <temma for>
|         <temma attr>
|         <temma if>
|           <temma last>
|       "x"

#data
<t:try>hoge<t:catch>fuga<t:catch>aaa</t:try><!---->
#errors
#document
| <!DOCTYPE html>
| <html>
|   <temma try>
|     "hoge"
|     <temma catch>
|     "fuga"
|     <temma catch>
|     "aaa"
|   <!--  -->

#data
<link><t:try>hoge<t:catch>fuga<t:catch>aaa</t:try><!---->
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <link>
|     <temma try>
|       "hoge"
|       <temma catch>
|       "fuga"
|       <temma catch>
|       "aaa"
|     <!--  -->

#data
<br><t:try>hoge<t:catch>fuga<t:catch>aaa</t:try><!---->
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <br>
|     <temma try>
|       "hoge"
|       <temma catch>
|       "fuga"
|       <temma catch>
|       "aaa"
|     <!--  -->

#data
<p><t:try>hoge</t:try>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <p>
|       <temma try>
|         "hoge"

#data
<t:catch>a
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <temma catch>
|     "a"

#data
<p><t:try>aa<a>fuga<t:catch>bb</t:try>aa
#errors
1;29;m;not closed;;a
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <p>
|       <temma try>
|         "aa"
|         <a>
|           "fuga"
|         <temma catch>
|         "bb"
|       "aa"

#data
<p><t:try>aaa<p>bbb</t:try>aaa
#errors
1;17;m;not closed;;t:try
1;28;m;unmatched end tag;;t:try
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <p>
|       <temma try>
|         "aaa"
|     <p>
|       "bbbaaa"

#data
<p><t:try>hh<t:try>aa<t:catch>b</t:try>cc<t:catch>dd</t:try>x
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <p>
|       <temma try>
|         "hh"
|         <temma try>
|           "aa"
|           <temma catch>
|           "b"
|         "cc"
|         <temma catch>
|         "dd"
|       "x"
